refactor(server): 优化登录失败次数的 Redis 存储类型- 将登录失败次数的存储类型从 Long 改为 Integer- 简化了类型转换过程，提高了代码的可读性和性能

+++ server/src/main/java/edp/davinci/controller/LoginController.java
@@ -116,7 +116,7 @@ public class LoginController {
             
             // 检查登录失败次数
             Object attemptsObj = redisUtils.get(attemptsKey);
-            int attempts = attemptsObj == null ? 0 : ((Long) attemptsObj).intValue();
+            int attempts = attemptsObj == null ? 0 : (Integer) attemptsObj;
             if (attempts >= maxAttempts) {
                 // 锁定账户5分钟
                 redisUtils.set(lockKey, true, (long) lockoutDuration, TimeUnit.SECONDS);
